+++ R/globals.R
@@ -7,5 +7,7 @@ utils::globalVariables(c("DATE", "geometry", "Station", "Values", "Variables",
                          "RAIN_YRS", "day", "elev", "lat", "lon", "rain_yrs", "wgn_id",
                          "max_value", "min_value", "lc1", "id", "lu", "type", "variables",
                          "PCP", "eq.label","rr.label", "SNAM", "TEXTURE", "param", "values",
-                         "NLAYERS", "elevation", "nbyr", "tstep"))
+                         "NLAYERS", "elevation", "nbyr", "tstep", "ANION_EXCL", "HYDGRP",
+                         "LAYER_NB", "PAR_NAME", "PAR_VAL", "SOL_CRK", "SOL_PH", "SOL_Z",
+                         "SOL_ZMX"))